requirements.txt
# --- Streamlit ---
streamlit

# --- LangChain Core & Integrations ---
langchain
langchain_community
langchain-openai
langchain-qdrant

# --- LangGraph ---
langgraph

# --- Vector Database Client ---
qdrant-client

# --- Document Loaders Dependencies ---
pypdf # For loading PDF files

# --- Environment Variables ---
python-dotenv